Add run-script generated Config
