+++ SuperheroSquadMaker/Sources/Config/Config.swift
@@ -1,0 +1,10 @@
+//! This file was generated by a RunScript Build Phase
+//! ⚠️ Do not modify or commit ⚠️
+
+import Foundation
+
+struct Config {
+	static let privateKey = "67a407d72945a9b1ee019145174ae42a02b7117d"
+	static let host = "gateway.marvel.com"
+	static let publicKey = "71e742407ca8b4ac76c4917a1c199a09"
+}
